[FOLD] Split implementations from interfaces: Moving implementation of Consensus and LedgerConsensus to ConsensusImp and LedgerConsensusImp respectively allows us to better levelize the code and reduce public interfaces.
